Move crud operations into "form-manager", along with DocumentImporter.
