+++ apps/spotlight/src/pages/forms/index.astro
@@ -1,5 +1,5 @@
 ---
-import { FormSection } from '@atj/design/src/form';
+import { FormSection } from '@atj/design/src/form-manager';
 import ContentLayout from '../../layouts/ContentLayout.astro';
 ---
