+++ apps/spotlight/src/pages/ud105-evicition-form.astro
@@ -1,5 +1,5 @@
 ---
-import { FormView } from '@atj/design/src/form/FormView';
+import { FormView } from '@atj/design/src/form-manager/FormView';
 import formData from '../htmlParser/ud105-form-field-output.json';
 import ContentLayout from '../layouts/ContentLayout.astro';
 ---
