+++ packages/design/src/form-manager/import-document.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createBrowserFormService } from '@atj/form-service';
 
-import DocumentImporter from '../form-builder/DocumentImporter';
+import DocumentImporter from './DocumentImporter';
 
 export const FormDocumentImport = ({ formId }: { formId: string }) => {
   const formService = createBrowserFormService();
