+++ packages/design/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
+    "baseUrl": "./src",
     "emitDeclarationOnly": true,
     "jsx": "react",
     "outDir": "./dist",
